Split codes to remove circluar import
