+++ main/Checkbox/CheckboxGroup.tsx
@@ -1,41 +1,9 @@
+import { CheckboxOptionType, CheckboxValueType, CustomStyle } from './types';
 import React, { FC, useCallback, useEffect, useRef, useState } from 'react';
 
 import { Checkbox } from './Checkbox';
+import { GroupCheckboxContext } from './GroupCheckboxContext';
 import styled from 'styled-components/native';
-
-interface CustomStyle {
-  labelSize?: number;
-  labelColor?: string;
-  boxSize?: number;
-  boxColor?: string;
-}
-
-interface OnChangeEvent {
-  checked: boolean;
-  label: string;
-}
-
-export type CheckboxValueType = string | number | boolean;
-
-export interface CheckboxOptionType {
-  label: string | number;
-  value: CheckboxValueType;
-  disabled?: boolean;
-  onChange?: (e: OnChangeEvent) => void;
-  customStyle?: CustomStyle;
-}
-
-export interface CheckboxGroupState {
-  value: CheckboxValueType[];
-  registeredValues: CheckboxValueType[];
-}
-
-export interface CheckboxGroupContext {
-  registerValue?: (value: CheckboxValueType) => void;
-  toggleOption: (option: CheckboxOptionType) => void;
-  value?: CheckboxValueType[];
-  disabled?: boolean;
-}
 
 interface CheckboxGroupProps {
   defaultValues?: Array<CheckboxValueType>;
@@ -46,8 +14,6 @@ interface CheckboxGroupProps {
   commonCustomStyle?: CustomStyle;
   direction?: 'column' | 'row';
 }
-
-export const GroupCheckboxContext = React.createContext<CheckboxGroupContext | null>(null);
 
 export const CheckboxGroup: FC<CheckboxGroupProps> = ({
   defaultValues,
