+++ main/Checkbox/GroupCheckboxContext.ts
@@ -1,0 +1,4 @@
+import { CheckboxGroupContext } from './types';
+import { createContext } from 'react';
+
+export const GroupCheckboxContext = createContext<CheckboxGroupContext | null>(null);
